% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{count_filter}
\alias{count_filter}
\title{Filters data by count}
\usage{
count_filter(x, greater_than_or_equal_to = NULL, less_than = NULL)
}
\arguments{
\item{x}{A data.frame containing a NUM column}

\item{greater_than_or_equal_to}{A number which indicates the minimum count
to include}

\item{less_than}{A number which indicates the maximum count
to include}
}
\value{
A data.frame subset by count
}
\description{
Subsets data by count
}

