% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abundance.R
\name{domain_abundance}
\alias{domain_abundance}
\title{Domain Abundance}
\usage{
domain_abundance(x, ntot)
}
\arguments{
\item{x}{A data.frame which is the output from the strat_density function}
}
\value{
A data.frame with abundance per sampling domain by species, its variance (var),
the number of SSUs per stratum (nm), the number of PSUs per stratum (n), the total
possible number of SSUs (NM), and the total possible number of PSUs (N)
}
\description{
Number of individuals in a sampling domain based on count. \strong{Note:} Does
not take into account detection probability.
}

