% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density.R
\name{domain_density}
\alias{domain_density}
\title{Calculates sampling domain level density per ssu}
\usage{
domain_density(x, ntot)
}
\arguments{
\item{x}{A data.frame including columns: REGION, YEAR, STRAT, PROT, PRIMARY_SAMPLE_UNIT,
STATION_NR, SPECIES_CD, NUM}

\item{ntot}{A data.frame including columns REGION, YEAR, STRAT, PROT, NTOT, GRID_SIZE}
}
\value{
A data.frame with density per SSU by species, its variance (var),
the number of SSUs in the domain (nm), the number of PSUs in the domain (n), the total
possible number of SSUs (NM), and the total possible number of PSUs (N)
}
\description{
Number of individuals per Secondary Sampling Unit (~177m^2) calculated
for each species at the stratum level
}

