% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbar.R
\name{domain_lbar}
\alias{domain_lbar}
\title{Domain mean length}
\usage{
domain_lbar(x, ntot)
}
\arguments{
\item{x}{A list containing SSU, PSU, and Stratum data, which is the output of \code{strat_lbar}.}

\item{ntot}{A data.frame of strata information.}
}
\value{
A data.frame with columns: Lbar (mean length), vbar_L (mean variance of Lbar), 
n (number of sampled primary sampling units), nm (number of sampled secondary sampling units), 
and STAGE_LEVEL.
}
\description{
Mean length (cm) at sampling domain level.
}
