% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/length_frequency.R
\name{domain_length_frequency}
\alias{domain_length_frequency}
\title{Domain level length frequency}
\usage{
domain_length_frequency(x, ntot)
}
\arguments{
\item{x}{A data.frame which is the output of strat_length_frequency}

\item{ntot}{A data.frame including columns: REGION, YEAR, STRAT, PROT, NTOT, GRID_SIZE}
}
\value{
A data.frame with: length frequencies for each stratum, the number of PSUs
per domain (n), the number of SSUs per domain (nm), the number of
possible PSUs per domain (N), and the number of possible SSUs per domain (NM)
}
\description{
Calculates the length frequecy at each sampling domain
}

