% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/richness.R
\name{domain_richness}
\alias{domain_richness}
\title{Species richness at the sampling domain level}
\usage{
domain_richness(x, ntot)
}
\value{
A data.frame with average species richness per SSU
for each sampling domain
}
\description{
Calculates the average number of species observed
per SSU for each sampling domain
}

