% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/total_biomass.R
\name{domain_total_biomass}
\alias{domain_total_biomass}
\title{Total sampling domain level biomass}
\usage{
domain_total_biomass(x, ntot)
}
\arguments{
\item{x}{A data.frame which is the output from psu_density}

\item{ntot}{A data.frame including columns: REGION, YEAR, STRAT, PROT, NTOT, GRID_SIZE}
}
\value{
A data.frame including total sampling domain biomass (total_biomass)
the average variance in total_biomass (var), number of PSUs sampled (n),
number of SSUs sampled (nm), number of possible PSUs (N), number of
possible SSUs (NM)
}
\description{
Total biomass on the domain level. Estimated from count
data and allomeric growth coefficients. \strong{Note:}
Does not take into account variance in growth parameters
or detection probability.
}

