% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.download_csv}
\alias{.download_csv}
\title{Download, unzip, and write a file to a csv}
\usage{
.download_csv(u, zipped)
}
\arguments{
\item{u}{A string. The URL at which to find the filehttps://grunt.sefsc.noaa.gov/rvc_analysis20/}

\item{zipped}{A boolean indiating whether a file is zipped file
or not}
}
\description{
Download, unzip, and write a file to a csv
}
