% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.getData}
\alias{.getData}
\title{A prototype for downloading csv files}
\usage{
.getData(years, regions, server, path, zipped, quiet)
}
\arguments{
\item{years}{A numeric vector of years for which to get sample data}

\item{regions}{A character vector of region codes for which to get sample data (e.g. "FLA KEYS",
"DRY TORT", "SEFCRI")}

\item{server}{A string containing the domain url at which to access the server}

\item{path}{A string of the relative path to this element of the api}

\item{zipped}{A boolean indiating whether a file is zipped file
or not}

\item{quiet}{A boolean indicating whether messages should be printed or not}
}
\description{
A prototype for downloading csv files
}
