% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrapppers.R
\name{getDomainLengthFrequency}
\alias{getDomainLengthFrequency}
\title{Sampling domain length frequency}
\usage{
getDomainLengthFrequency(x, species, length_bins = NULL,
  merge_protected = TRUE, ...)
}
\value{
A data.frame with length frequencies for each sampling domain
}
\description{
Calculates sampling domain level frequencies at length
}

