% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{getRvcData}
\alias{getRvcData}
\title{Downloads RVC data and saves to a list}
\usage{
getRvcData(years, regions, server = "http://localhost:3000")
}
\arguments{
\item{years}{A numeric vector of years for which to get sample data}

\item{regions}{A character vector of region codes for which to get sample data (e.g. "FLA KEYS",
"DRTO", "SEFCRI")}

\item{server}{A string containing the domain url at which to access the server}
}
\value{
A list containing three data frames: sample_data - containing
sample information, stratum_data - containing stratum information,
taxonomic_data - containing taxonomic and life history information
}
\description{
Dowloads sample, stratum, and taxonomic/life-history data for
provided years and regions, and saves them to a list which can be used in
summary statistic procuding functions in the package
}
\details{
To understand more about the meaning of the different variables
in the returned list please see the documentation for
\code{\link{getSampleData}}, \code{\link{getStratumData}},
\code{\link{getTaxonomicData}}
}
\seealso{
\code{\link{getStratumData}} \code{\link{getTaxonomicData}} \code{\link{getSampleData}}
}

