% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density_wrappers.R
\name{getStratumDensity}
\alias{getStratumDensity}
\title{Stratum level density}
\usage{
getStratumDensity(x, species, length_bins = NULL, merge_protected = TRUE,
  ...)
}
\arguments{
\item{x}{A list containing three data.frames:
\itemize{
 \item{sample_data: }{sample
 counts and information}
 \item{stratum_data: }{strata information. Including the number
 of possible primary sampling units per stratum (NTOT)}
 \item{taxonomic_data: }{taxonomic and life history information}
}}

\item{species}{A character vector containing scientific names, common names, or species codes for the desired species,
or a combination thereof}

\item{length_bins}{Numeric vector, data.frame, or keyword. \itemize{
\item{vector: }{A number of numeric vector of lengths, in cm, of breakpoints by which the data will
be binned. The same vector will be applied to all species. \strong{NOTE:}{ Be wary of estimates produced
using small length bins, as the associated variance may be large or misleading}}
\item{data.frame: }{A lookup table containing two columns. The first column with scientific names
 or species codes, and the second containing lengths, in cm, by which to break up the data for each species}
 \item{keyword: }{"lc" - breaks the data at minimum length at capture for each species.
 "lm" - breaks the data at median length at maturity for each species. \strong{NOTE:} Generates
 length at capture and length at maturity from taxonomic data table. Can only be used if
 length at capture or length at maturity is available for the species in the taxonomic_data table. See
 \code{\link{getTaxonomicData}}}
}}

\item{merge_protected}{A boolean indicating whether protected and unprotected areas should be merged (TRUE, the default),
or should be calculated seperately (FALSE)}

\item{...}{Optional filters to apply to the data:
\describe{
 \item{strata}{Character vector of strata codes to include. Strata codes vary by region}
 \item{status}{Numeric vector of protected statuses}
 \item{is_protected}{Boolean indicating whether only protected areas should be included (TRUE),
 only unprotected areas (FALSE), or both (NULL, the default). Must be NULL if merge_protected is FALSE}
 \item{years}{Numeric vector of years to include}
 \item{regions}{Character vector of region codes: (e.g. "FLA KEYS", "DRY TORT", "SEFCRI") to include}
 \item{when_present}{Boolean indicating whether to only include records where individuals present (TRUE),
 or not (FALSE)}
 \item{group}{A lookup table (data.frame): the first column of which is a list of species codes,
 scientific names, or common names, and the second column of which is a list of names by which to
 group the species (e.g. Family names, trophic groups, etc). If this option is used instead of the
 statistic being calculated per species, it will be calculated per group}
}}
}
\value{
A data.frame with:
\describe{
\item{YEAR}{The year}
\item{REGION}{A code for the region: DRY TORT - Dry Tortugas, SEFCRI - Southeast Peninsular Florida,
FLA KEYS - Florida Keys}
\item{STRAT}{A code for the stratum}
\item{PROT}{A boolean indicating protected status: 1 - Protected, 2 - Unprotected}
\item{SPECIES_CD}{The species code. The first three letters of the genus name and first four
of the species name. If group is passed as an argument, SPECIES_CD will be changed to GROUP}
\item{density}{Average density per secondary sampling unit}
\item{var}{Variance in average density per secondary sampling unit}
\item{n}{Number of primary sampling units sampled}
\item{nm}{Number of secondary sampling units sampled}
\item{N}{Number of possible primary sample units}
\item{NM}{Number of possible secondary sampling units}
\item{length_class}{The length class or bin. Only present if length_bins is not NULL.
The notation, [lower, upper), is inclusive of the lower bound, but exclusive of the upper bound}
\item{protected_status}{The protected status. Only present if merge_protected is FALSE}
}
}
\description{
Calculates average density (individual/secondary sampling unit) for each
stratum
}
\examples{
## Get RVC data from 2012 in the Florida Keys
fk2012 = getRvcData(years = 2011, regions = "FLA KEYS")

## Calculate stratum density for Red Grouper
getStratumDensity(fk2012, species = "Red Grouper")
}
\seealso{
\code{\link{getRvcData}} \code{\link{getDomainDensity}}
}

