% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{length_filter}
\alias{length_filter}
\title{Filters data by length}
\usage{
length_filter(x, len_geq = NULL, len_lt = NULL, ...)
}
\arguments{
\item{x}{A data.frame containing a LEN and NUM column}

\item{len_geq}{A number which indicates the minimum length
in centimeters (inclusive).
NULL (default) will return the original data}

\item{len_lt}{A number which indicates the maximum length
in centimeters (exclusive).
NULL (default) will return the original data}

\item{...}{Anonymous arguments to be passed to other filters}
}
\value{
A data.frame with the original data subset by length
}
\description{
Sets count to zero for records where length is not within specified parameters
}

