% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{protected_filter}
\alias{protected_filter}
\title{Filter data by protected status}
\usage{
protected_filter(x, status = NULL, is_protected = NULL)
}
\arguments{
\item{x}{A data.frame containing a PROT column}

\item{status}{A numeric vector specifying the protected status.
Generally, 0 for unprotected, can be 1 or 2 for
different levels of protection. Defuault value of NULL
will return all protected statuses}

\item{is_protected}{A boolean indicating whether to select only
protected (TRUE), unprotected (FALSE), or both (NULL, the default)}
}
\description{
Subsets data by protected status
}

