% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biomass.R
\name{psu_biomass}
\alias{psu_biomass}
\title{PSU biomass}
\usage{
psu_biomass(x)
}
\arguments{
\item{x}{A data.frame which is the output of ssu_biomass}
}
\value{
A data.frame with the average biomass per secondary sampling unit
for each primary sampling unit, and the variance thereof
}
\description{
Biomass (in kg) per secondary sampling unit for each
primary sampling unit
}

