% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density.R
\name{psu_density}
\alias{psu_density}
\title{Calculates PSU level density per SSU}
\usage{
psu_density(x)
}
\arguments{
\item{x}{A data.frame including columns: REGION, YEAR, STRAT, PROT, PRIMARY_SAMPLE_UNIT,
STATION_NR, SPECIES_CD, NUM}
}
\value{
A data.frame with density per SSU aggregated by PSU and species,
variance in density (var), and the number of ssus per psu (m)
}
\description{
Number of individuals per Secondary Sampling Unit (~177m^2) calculated
for each species at the primary sampling unit level
}

