% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/occurence.R
\name{psu_occurrence}
\alias{psu_occurrence}
\title{PSU Occurrence}
\usage{
psu_occurrence(x)
}
\arguments{
\item{x}{A data.frame which is the output of ssu_occurrence}
}
\value{
A data.frame with a column, occurrence, of the average occurrence
per secondary sampling unit (~177m^2) for a primary sampling unit,
its associated between SSU variance (var), and the number of secondary sampling units
sampled per primary sampling unit (m)
}
\description{
Average occurrence per secondary sampling unit at the primary sampling unit level
}

