% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/richness.R
\name{psu_richness}
\alias{psu_richness}
\title{Species richness at the PSU level}
\usage{
psu_richness(x)
}
\arguments{
\item{x}{A data.frame including columns: REGION, YEAR, STRAT, PROT, PRIMARY_SAMPLE_UNIT,
STATION_NR, SPECIES_CD, NUM}
}
\value{
A data.frame with average species richness per SSU for
each PSU
}
\description{
Calculates the average number of species observed
per SSU for each PSU
}

