% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{region_filter}
\alias{region_filter}
\title{Filters data by region}
\usage{
region_filter(x, regions = NULL)
}
\arguments{
\item{x}{A data.frame containing a region column}

\item{regions}{A character vector of regions codes. NULL (default)
returns original data}
}
\value{
A data.frame of the original data
subset by region
}
\description{
Subsets data by region
}

