% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biomass.R
\name{ssu_biomass}
\alias{ssu_biomass}
\title{SSU Biomass}
\usage{
ssu_biomass(x, growth_parameters)
}
\arguments{
\item{x}{A data.frame containing columns: YEAR, REGION, STRAT, PROT, PRIMARY_SAMPLE_UNIT,
STATION_NR, SPECIES_CD, NUM, LEN}

\item{growth_parameters}{A list of allometric growth parameters, containing: a - the linear coefficient in kg/cm, and b - the
exponential coefficient.}
}
\value{
A data.frame containing a column biomass with the biomass per secondary sampling unit
}
\description{
Biomass (in kg) per secondary sampling unit
}
\details{
The form of the allometric growth equation used in calculating biomass is:
\deqn{
 W(kg) = a(kg/cm)L(cm)^b
}
}

