% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density.R
\name{ssu_density}
\alias{ssu_density}
\title{Calculates SSU level density per ssu}
\usage{
ssu_density(x)
}
\arguments{
\item{x}{A data.frame including columns REGION, YEAR, STRAT, PROT, PRIMARY_SAMPLE_UNIT,
STATION_NR, SPECIES_CD, NUM, and LEN (if by_length = TRUE)}
}
\value{
A data.frame with counts aggregated by SSU, SPECIES_CD
}
\description{
Sums the count per species at the SSU level
}

