% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/occurence.R
\name{ssu_occurrence}
\alias{ssu_occurrence}
\title{SSU Occurrence}
\usage{
ssu_occurrence(x)
}
\arguments{
\item{x}{A data.frame including columns: REGION, YEAR, STRAT, PROT, PRIMARY_SAMPLE_UNIT,
STATION_NR, SPECIES_CD, NUM}
}
\value{
A data.frame with a column, occurrence, indicating whether or not
a species was present in a particular secondary sampling unit
}
\description{
Whether or not a particular species is present in a secondary sampling unit
}

