% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/richness.R
\name{ssu_richness}
\alias{ssu_richness}
\title{Species richness at the SSU level}
\usage{
ssu_richness(x)
}
\arguments{
\item{x}{A data.frame including columns: REGION, YEAR, STRAT, PROT, PRIMARY_SAMPLE_UNIT,
STATION_NR, SPECIES_CD, NUM}
}
\value{
A data.frame with species richness per SSU
}
\description{
Calculates the number of species observed at the
secondary sampling unit level
}

