% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abundance.R
\name{strat_abundance}
\alias{strat_abundance}
\title{Stratum abundance}
\usage{
strat_abundance(x, ntot)
}
\arguments{
\item{x}{A data.frame which is the output from psu_density}

\item{ntot}{A data.frame including columns: REGION, YEAR, STRAT, PROT, NTOT, GRID_SIZE}
}
\value{
A data.frame with abundance per stratum by species, its variance (var),
the number of SSUs per stratum (nm), the number of PSUs per stratum (n), the total
possible number of SSUs (NM), and the total possible number of PSUs (N)
}
\description{
Number of individuals in a stratum based on count. \strong{Note:} Does
not take into account detection probability.
}

