% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biomass.R
\name{strat_biomass}
\alias{strat_biomass}
\title{Stratum level biomass}
\usage{
strat_biomass(x, ntot)
}
\arguments{
\item{x}{A data.frame which is the output of psu_biomass}
}
\value{
A data.frame with the average biomass per secondary sampling unit for each stratum,
its variance (var), the number of PSUs sampled (n), the number of SSUs sampled (nm),
the number of possible PSUs (N), and the number of possible SSUs (NM)
}
\description{
Biomass (in kg) per secondary sampling unit for each
stratum
}

