% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density.R
\name{strat_density}
\alias{strat_density}
\title{Calculates the Stratum level density per SSU}
\usage{
strat_density(x, ntot)
}
\arguments{
\item{x}{A data.frame including columns REGION, YEAR, STRAT, PROT, SPECIES_CD, NUM}

\item{ntot}{A data.frame including columns REGION, YEAR, STRAT, PROT, NTOT, GRID_SIZE}
}
\value{
A data.frame with density per SSU aggregated by stratum, its variance (var),
the number of SSUs per stratum (nm), the number of PSUs per stratum (n), the total
possible number of SSUs (NM), and the total possible number of PSUs (N)
}
\description{
Calculates the Stratum level density per SSU, its variance (var), the number of SSUs per
stratum (nm), and the number of PSUs per stratum (n)
}

