% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbar.R
\name{strat_lbar}
\alias{strat_lbar}
\title{Stratum mean length}
\usage{
strat_lbar(x, ntot)
}
\arguments{
\item{x}{A list containing SSU and PSU data, which is the output of \code{psu_lbar}.}

\item{ntot}{A data.frame of strata information.}
}
\value{
A list containing three data.frames: 
\itemize{
  \item \code{strat_dat}: Stratum-level summaries including mean sum of counts (\code{xi}), mean sum of lengths (\code{yi}), and sampling unit counts (\code{n}, \code{nm}).
  \item \code{psu_dat}: The PSU-level data passed from the previous step.
  \item \code{ssu_dat}: The SSU-level data passed from the previous step.
}
}
\description{
Mean length at the stratum level.
}
