% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/length_frequency.R
\name{strat_length_frequency}
\alias{strat_length_frequency}
\title{Stratum level lenth frequencies}
\usage{
strat_length_frequency(x, ntot)
}
\arguments{
\item{x}{A data.frame containing columns: YEAR, REGION, STRAT, PROT, PRIMARY_SAMPLE_UNIT,
STATION_NR, NUM, and LEN}
}
\value{
A data.frame with length frequencies for each stratum
}
\description{
Calculates length frequency at each stratum
}
\details{
The LEN column should be a numeric column, but the lengths need not
be the original measured lengths. If they are binned, the LEN column
should be transfomed to represent the midpoint of each bin.
}

