% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/occurence.R
\name{strat_occurrence}
\alias{strat_occurrence}
\title{Stratum occurrence}
\usage{
strat_occurrence(x, ntot)
}
\arguments{
\item{x}{A data.frame which is the output of psu_occurrence}

\item{ntot}{A data.frame including columns: REGION, YEAR, STRAT, PROT, NTOT, GRID_SIZE}
}
\value{
A data.frame with columns: occurrence, the average occurrence per SSU; var, the
average variance in occurrence; n, the number of sampled primary sampling units;
nm, the number of sampled secondary sampling units; N, the number of possible
primary sampling units; NM, the number of possible secondary sampling units
}
\description{
Average occurrence per secondary sampling unit at the stratum level
}

