% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/richness.R
\name{strat_richness}
\alias{strat_richness}
\title{Species richness at the stratum level}
\usage{
strat_richness(x, ntot)
}
\arguments{
\item{x}{A data.frame which is the output from psu_density}

\item{ntot}{A data.frame including columns: REGION, YEAR, STRAT, PROT, NTOT, GRID_SIZE}
}
\value{
A data.frame with average species richness per SSU
for each stratum
}
\description{
Calculates the average number of species observed
per SSU for each stratum
}

