% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/total_biomass.R
\name{strat_total_biomass}
\alias{strat_total_biomass}
\title{Total stratum level biomass}
\usage{
strat_total_biomass(x, ntot)
}
\arguments{
\item{x}{A data.frame which is the output of psu_biomass}
}
\value{
A data.frame including total stratum biomass (total_biomass),
the average variance in total_biomass (var), number of PSUs sampled (n),
number of SSUs sampled (nm), number of possible PSUs (N), number of
possible SSUs (NM)
}
\description{
Total biomass on the stratum level. Estimated from count
data and allometric growth coefficients. \strong{Note:}
Does not take into account variance in growth parameters
or detection probability
}

