% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{strata_filter}
\alias{strata_filter}
\title{Filter data by stratum}
\usage{
strata_filter(x, strata = NULL)
}
\arguments{
\item{x}{A data.frame containing a STRAT column}

\item{strata}{A character vector containing the stratum codes by which to subset the data.
NULL (default) will return the original data}
}
\value{
A data.frame of the original data subset by
strata
}
\description{
Subsets data by species
}

