% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{year_filter}
\alias{year_filter}
\title{Filters data by year}
\usage{
year_filter(x, years = NULL)
}
\arguments{
\item{x}{A data.frame containing a YEAR column}

\item{years}{A numeric vector of years. NULL (defualt)
returns original data}
}
\value{
A data.frame of the original data
subset by years
}
\description{
Subsets data by year
}

