% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{count_filter}
\alias{count_filter}
\title{Filters data by count}
\usage{
count_filter(x, cnt_geq = NULL, cnt_lt = NULL, when_present = FALSE, ...)
}
\arguments{
\item{x}{A data.frame containing a NUM column}

\item{cnt_geq}{A number which indicates the minimum count
to include}

\item{cnt_lt}{A number which indicates the maximum count
to include}

\item{when_present}{A boolean. Selects samples where NUM > 0 if TRUE. False (default)
does not subset data.}

\item{...}{Anonymous arguments to be passed to other filters}
}
\value{
A data.frame subset by count
}
\description{
Subsets data by count
}

