% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getDomainLengthFrequency}
\alias{getDomainLengthFrequency}
\title{Domain level length frequency}
\usage{
getDomainLengthFrequency(x, species, length_bins = NULL,
  merge_protected = TRUE, ...)
}
\arguments{
\item{x}{A list containing three data.frames:
\itemize{
 \item{sample_data: }{sample
 counts and information}
 \item{stratum_data: }{strata information. Including the number
 of possible primary sampling units per stratum (NTOT)}
 \item{taxonomic_data: }{taxonomic and life history information}
}}

\item{species}{A character vector containing scientific names, common names, or species codes for the desired species,
or a combination thereof}

\item{length_bins}{A numeric vector of lengths, in centimenters, by which to bin the data.
If NULL (default), will not bin the data.
\strong{Note:} Lengths below the minimum and above the maximum
value in length_bins will be assigned <NA> (with no differentiation between
below and above the provided bins). It is suggested that
the bins match or exceed the range of lengths for the species in which you
are interested}

\item{merge_protected}{A boolean indicating whether protected and unprotected areas should be merged (TRUE, the default),
or should be calculated seperately (FALSE)}

\item{...}{Optional filters to apply to the data:
\describe{
 \item{strata}{Character vector of strata codes to include. Strata codes vary by region}
 \item{status}{Numeric vector of protected statuses}
 \item{is_protected}{Boolean indicating whether only protected areas should be included (TRUE),
 only unprotected areas (FALSE), or both (NULL, the default). Must be NULL if merge_protected is FALSE}
 \item{years}{Numeric vector of years to include}
 \item{regions}{Character vector of region codes: (e.g. "FLA KEYS", "DRY TORT", "SEFCRI") to include}
 \item{when_present}{Boolean indicating whether to only include records where individuals present (TRUE),
 or not (FALSE)}
}}
}
\value{
A data.frame with:
\describe{
\item{YEAR}{The year}
\item{REGION}{A code for the region: DRY TORT - Dry Tortugas, SEFCRI - Southeast Peninsular Florida,
FLA KEYS - Florida Keys}
\item{SPECIES_CD}{The species code. The first three letters of the genus name and first four
of the species name}
\item{length_class}{The actual measured lengths, in cm, if length_bins = NULL, or an interval, if length_bins
is not NULL}
\item{frequency}{The relative frequency of each length class within each sampling domain}
}
}
\description{
Calculates length frequency for each sampling domain
}
\examples{
## Get rvc data from the Florida Keys in 2005
fk2005 = getRvcData(years = 2005, regions = "FLA KEYS")

## Calculate stratum length frequencies for Mangrove Snapper
getDomainLengthFrequency(fk2005, species = "LUT GRIS")

## Calculate length frequencies for mangrove snappers in
## 1cm bins
getDomainLengthFrequency(fk2005, species = "LUT GRIS", length_bins = seq(0,100,1))
}
\seealso{
\code{\link{getRvcData}} \code{\link{getStratumLengthFrequency}}
}

