% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_wrappers.R
\name{getDomainTotalBiomass}
\alias{getDomainTotalBiomass}
\title{Domain level total biomass}
\usage{
getDomainTotalBiomass(
  x,
  species,
  growth_parameters = NULL,
  length_bins = NULL,
  merge_protected = TRUE,
  ...
)
}
\arguments{
\item{x}{A list containing three data.frames:
\itemize{
 \item{sample_data: }{sample
 counts and information}
 \item{stratum_data: }{strata information. Including the number
 of possible primary sampling units per stratum (NTOT)}
 \item{taxonomic_data: }{taxonomic and life history information}
}}

\item{species}{A character vector containing scientific names, common names, or species codes for the desired species,
or a combination thereof}

\item{growth_parameters}{A data.frame or list of allometric growth parameters.\itemize{
\item{list: }{A list with variables named 'a' and 'b'
representing the linear and exponential coefficient of the allometric growth equation (see details)}
\item{data.frame: }{A data.frame with columns named a and b (or WLEN_A and WLEN_B) as well as a SPECIES_CD
column. If growth_parameters is NULL (default) function will attempt to use parameters from taxonomic data
retrieved from server}}
\strong{NOTE:} using a list means that only one set of growth parameters will be used for all species,
while passing in a data.frame will allow the function to lookup the growth_parameters for each species}

\item{length_bins}{Numeric vector, data.frame, or keyword. \itemize{
\item{vector: }{A number of numeric vector of lengths, in cm, of breakpoints by which the data will
be binned. The same vector will be applied to all species. \strong{NOTE:}{ Be wary of estimates produced
using small length bins, as the associated variance may be large or misleading}}
\item{data.frame: }{A lookup table containing two columns. The first column with scientific names
 or species codes, and the second containing lengths, in cm, by which to break up the data for each species}
 \item{keyword: }{"lc" - breaks the data at minimum length at capture for each species.
 "lm" - breaks the data at median length at maturity for each species. \strong{NOTE:} Generates
 length at capture and length at maturity from taxonomic data table. Can only be used if
 length at capture or length at maturity is available for the species in the taxonomic_data table. See
 \code{\link{getTaxonomicData}}}
}}

\item{merge_protected}{A boolean indicating whether protected and unprotected areas should be merged (TRUE, the default),
or should be calculated seperately (FALSE)}

\item{...}{Optional filters to apply to the data:
\describe{
 \item{strata}{Character vector of strata codes to include. Strata codes vary by region}
 \item{status}{Numeric vector of protected statuses}
 \item{is_protected}{Boolean indicating whether only protected areas should be included (TRUE),
 only unprotected areas (FALSE), or both (NULL, the default). Must be NULL if merge_protected is FALSE}
 \item{years}{Numeric vector of years to include}
 \item{regions}{Character vector of region codes: (e.g. "FLA KEYS", "DRY TORT", "SEFCRI") to include}
 \item{when_present}{Boolean indicating whether to only include records where individuals present (TRUE),
 or not (FALSE)}
 \item{group}{A lookup table (data.frame): the first column of which is a list of species codes,
 scientific names, or common names, and the second column of which is a list of names by which to
 group the species (e.g. Family names, trophic groups, etc). If this option is used instead of the
 statistic being calculated per species, it will be calculated per group}
}}
}
\value{
A data.frame with:
\describe{
\item{YEAR}{The year}
\item{REGION}{A code for the region: DRY TORT - Dry Tortugas, SEFCRI - Southeast Peninsular Florida,
FLA KEYS - Florida Keys}
\item{SPECIES_CD}{The species code. The first three letters of the genus name and first four
of the species name. If group is passed as an argument, SPECIES_CD will be changed to GROUP}
\item{total_biomass}{Total biomass for each sampling domain}
\item{var}{Variance in total biomass for each sampling domain}
\item{n}{Number of primary sampling units sampled}
\item{nm}{Number of secondary sampling units sampled}
\item{N}{Number of possible primary sample units}
\item{NM}{Number of possible secondary sampling units}
\item{length_class}{The length class or bin. Only present if length_bins is not NULL.
The notation, [lower, upper), is inclusive of the lower bound, but exclusive of the upper bound}
\item{protected_status}{The protected status. Only present if merge_protected is FALSE}
}
}
\description{
Calculates total biomass for each sampling domain
}
\details{
The form of the allometric growth equation used in calculating biomass is:
\deqn{
  W[kg] = (a*(L[cm]*10[mm/cm])^b)/1000[kg/g]
}
The 'a' and 'b' parameters provided in growth_parameters should be for
converting from millimeters to grams, even though the final output is in kilograms.
}
\examples{
## Get RVC data for the Florida Keys in 2012
fk2012 = getRvcData(years = 2012, regions = "FLA KEYS")

## Calculate total biomass of Yellowtail Snapper for each stratum
getDomainTotalBiomass(fk2012, species = "Ocy chry", growth_parameters = list(a = 7.75e-5, b = 2.718))
}
\seealso{
\code{\link{getRvcData}} \code{\link{getStratumTotalBiomass}} \code{\link{getDomainBiomass}}
}
