% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrapppers.R
\name{getStratumAbundance}
\alias{getStratumAbundance}
\title{Stratum level abundance}
\usage{
getStratumAbundance(x, species, length_bins = NULL, ...)
}
\arguments{
\item{x}{An list containing three data.frames:
\itemize{
 \item{sample_data: }{sample
 counts and information}
 \item{stratum_data: }{strata information. Including the number
 of possible primary sampling units per stratum (NTOT)}
 \item{taxonomic_data: }{taxonomic and life history information}
}}

\item{species}{A character vector containing scientific names, common names, or species codes for the desired species,
or a combination thereof}

\item{length_bins}{A numeric vector of lengths, in centimenters, by which to bin the data. If NULL (default), will not bin the data}

\item{...}{Optional filters to apply to the data:
\describe{
 \item{strata}{Character vector of strata codes}
 \item{status}{Numeric vector of protected statuses.}
 \item{is_protected}{Boolean indicating whether only protected areas should be included (TRUE),
 only unportected areas (FALSE), or both (NULL, the default)}
 \item{years}{Numeric vector of years}
 \item{regions}{Character vector of region codes: (e.g. "FLA KEYS", "DRTO", "SEFCRI")}
 \item{when_present}{Boolean indicating whether to only include records where individuals present (TRUE),
 or not (FALSE). \strong{NOTE:} Using this option with multiple species will result in including samples
 where any of the species are present}
}}
}
\value{
A data.frame with the abundance for each stratum
}
\description{
Calculates stratum level abundance
}

