% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{getStratumData}
\alias{getStratumData}
\title{Download stratum data from server}
\usage{
getStratumData(years, regions, server = "http://localhost:3000")
}
\arguments{
\item{years}{A numeric vector of years for which to get sample data}

\item{regions}{A character vector of region codes for which to get sample data (e.g. "FLA KEYS",
"DRTO", "SEFCRI")}

\item{server}{A string containing the domain url at which to access the server}
}
\value{
A data.frame with stratum data for years and regions provided
}
\description{
Download stratum data from server and save as data.frame
}

