% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{getTaxonomicData}
\alias{getTaxonomicData}
\title{Download taxonomic data from server}
\usage{
getTaxonomicData(server = "http://localhost:3000")
}
\arguments{
\item{server}{A string containing the domain url at which to access the server}
}
\value{
A data.frame with taxonomic data and life history data for
all species in the RVC
}
\description{
Download taxonomic and life history data from server
}

