% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{length_filter}
\alias{length_filter}
\title{Filters data by length}
\usage{
length_filter(x, greater_than_or_equal_to = NULL, less_than = NULL)
}
\arguments{
\item{x}{A data.frame containing a LEN column}

\item{greater_than_or_equal_to}{A number which indicates the minimum length
in centimeters (inclusive) by which to subset the data.
NULL (default) will return the original data}

\item{less_than}{A number which indicates the maximum length
in centimeters (exclusive) by which to subset the data.
NULL (default) will return the original data}
}
\value{
A data.frame with the original data subset by length
}
\description{
Subsets data by length
}

