% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density.R
\name{psu_density}
\alias{psu_density}
\title{Calculates PSU level density per SSU}
\usage{
psu_density(x)
}
\arguments{
\item{x}{A data.frame including columns REGION, YEAR, STRAT, PROT, PRIMARY_SAMPLE_UNIT,
SPECIES_CD, NUM}
}
\value{
A data.frame with density per SSU aggregated by PSU, SPECIES_CD
variance in density (var), and the number of ssus per psu (m)
}
\description{
Calculates the PSU level density per SSU, its variance (var), and the number
of SSUs per PSU (m)
}

