% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{species_filter}
\alias{species_filter}
\title{Filter data by species}
\usage{
species_filter(x, species_cd = NULL)
}
\arguments{
\item{x}{A data.frame containing a SPECIES_CD column}

\item{species_cd}{A character vector containing the species codes by which to
subset data. NULL (default) will return the original data.
 \strong{Note:} Species codes contain the
first three letters of the genus name, a space, and the
first four letters of the species name.}
}
\value{
A data.frame of the original data subset by
species code
}
\description{
Subsets data by species
}

